Refactor render pipeline processing
